+++ RealmSwift-swift2.2/Tests/RealmTests.swift
@@ -590,14 +590,73 @@ class RealmTests: TestCase {
     }
 
     func testDynamicObjectForPrimaryKey() {
+        let intTypes: [Object.Type] = [SwiftPrimaryIntObject.self,
+                                       SwiftPrimaryInt8Object.self,
+                                       SwiftPrimaryInt16Object.self,
+                                       SwiftPrimaryInt32Object.self,
+                                       SwiftPrimaryInt64Object.self]
+        let optionalIntTypes: [Object.Type] = [SwiftPrimaryOptionalIntObject.self,
+                                               SwiftPrimaryOptionalInt8Object.self,
+                                               SwiftPrimaryOptionalInt16Object.self,
+                                               SwiftPrimaryOptionalInt32Object.self,
+                                               SwiftPrimaryOptionalInt64Object.self]
+
         let realm = try! Realm()
         try! realm.write {
             realm.create(SwiftPrimaryStringObject.self, value: ["a", 1])
             realm.create(SwiftPrimaryStringObject.self, value: ["b", 2])
+
+            realm.create(SwiftPrimaryOptionalStringObject.self, value: [NSNull(), 1])
+            realm.create(SwiftPrimaryOptionalStringObject.self, value: ["b", 2])
+
+            func createIntObject(objectType: Object.Type) {
+                realm.create(objectType, value: ["a", 1])
+                realm.create(objectType, value: ["b", 2])
+            }
+
+            func createOptionalIntObject(objectType: Object.Type) {
+                realm.create(objectType, value: ["a", NSNull()])
+                realm.create(objectType, value: ["b", 2])
+            }
+
+            for type in intTypes {
+                createIntObject(type)
+            }
+
+            for type in optionalIntTypes {
+                createOptionalIntObject(type)
+            }
         }
 
-        XCTAssertNotNil(realm.dynamicObjectForPrimaryKey("SwiftPrimaryStringObject", key: "a"))
-        XCTAssertNil(realm.dynamicObjectForPrimaryKey("SwiftPrimaryStringObject", key: "z"))
+        XCTAssertNotNil(realm.dynamicObjectForPrimaryKey(String(SwiftPrimaryStringObject), key: "a"))
+        XCTAssertNil(realm.dynamicObjectForPrimaryKey(String(SwiftPrimaryStringObject), key: "z"))
+
+        XCTAssertNotNil(realm.dynamicObjectForPrimaryKey(String(SwiftPrimaryOptionalStringObject), key: NSNull()))
+        XCTAssertEqual(realm.dynamicObjectForPrimaryKey(String(SwiftPrimaryOptionalStringObject), key: NSNull()),
+                       realm.dynamicObjectForPrimaryKey(String(SwiftPrimaryOptionalStringObject), key: nil))
+        XCTAssertNotNil(realm.dynamicObjectForPrimaryKey(String(SwiftPrimaryOptionalStringObject), key: "b"))
+        XCTAssertNil(realm.dynamicObjectForPrimaryKey(String(SwiftPrimaryOptionalStringObject), key: "z"))
+
+        func assertIntObject(objectType: Object.Type) {
+            XCTAssertNotNil(realm.dynamicObjectForPrimaryKey(String(objectType), key: 1))
+            XCTAssertNil(realm.dynamicObjectForPrimaryKey(String(objectType), key: 0))
+        }
+
+        func assertOptionalIntObject(objectType: Object.Type) {
+            XCTAssertNotNil(realm.dynamicObjectForPrimaryKey(String(objectType), key: NSNull()))
+            XCTAssertEqual(realm.dynamicObjectForPrimaryKey(String(objectType), key: NSNull()),
+                           realm.dynamicObjectForPrimaryKey(String(objectType), key: nil))
+            XCTAssertNotNil(realm.dynamicObjectForPrimaryKey(String(objectType), key: 2))
+            XCTAssertNil(realm.dynamicObjectForPrimaryKey(String(objectType), key: 0))
+        }
+
+        for type in intTypes {
+            assertIntObject(type)
+        }
+
+        for type in optionalIntTypes {
+            assertOptionalIntObject(type)
+        }
     }
 
     func testDynamicObjectForPrimaryKeySubscripting() {