+++ CHANGELOG.md
@@ -3,8 +3,7 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### API breaking changes
 
-* The `key` parameter of `objectForPrimaryKey(_:, key:)` is 
-  now marked as optional.
+* None.
 
 ### Enhancements
 
@@ -15,6 +14,8 @@ x.x.x Release notes (yyyy-MM-dd)
 * Swift: Fix an error thrown when trying to create or update `Object` instances via
   `add(:_update:)` with a primary key property of type `RealmOptional`.
 * Xcode playground in Swift release zip now runs successfully.
+* The `key` parameter of `Realm.objectForPrimaryKey(_:key:)`/ `Realm.dynamicObjectForPrimaryKey(_:key:)`
+ is now marked as optional.
 
 1.0.0 Release notes (2016-05-25)
 =============================================================